refactoring: controllers

+++ backend/app/api/routes/items/items_service.py
@@ -1,29 +1,35 @@
 import uuid
 from typing import Any
 
+from app.core.models.users import User
 from app.core.repositories.items import ItemsRepository
-from fastapi import APIRouter
-from sqlmodel import func, select
+from fastapi import APIRouter, HTTPException
+from sqlmodel import Session, func, select
 
-from app.api.deps import CurrentUser, SessionDep
 from app.core.models.items import Item
 from app.core.dtos.items import ItemCreate, ItemsPublic, ItemUpdate
 from app.core.dtos.message import Message
 
 router = APIRouter(prefix="/items", tags=["items"])
-    
+
+class ItemsValidationService:
+    @staticmethod
+    def check_current_user_permissions(item: Item, current_user: User):
+        if not item:
+            raise HTTPException(status_code=404, detail="Item not found")
+        if not current_user.is_superuser and (item.owner_id != current_user.id):
+            raise HTTPException(status_code=400, detail="Not enough permissions")
 
 class ItemsService:
     """Handles items-related business logic and database operations"""
 
     @staticmethod
     def read_items(
-        session: SessionDep, current_user: CurrentUser, skip: int = 0, limit: int = 100
+        session: Session, current_user: User, skip: int = 0, limit: int = 100
     ) -> Any:
         """
         Retrieve items.
         """
-
         if current_user.is_superuser:
             count_statement = select(func.count()).select_from(Item)
             count = session.exec(count_statement).one()
@@ -46,17 +52,19 @@ class ItemsService:
         return ItemsPublic(data=items, count=count)
 
     @staticmethod
-    def read_item(session: SessionDep, current_user: CurrentUser, id: uuid.UUID) -> Any:
+    def read_item(session: Session, current_user: User, id: uuid.UUID) -> Any:
         """
         Get item by ID.
         """
-        return ItemsRepository.get_user_item_by_id(
+        item = ItemsRepository.get_item_by_id(
             session, current_user, id,
         )
+        ItemsValidationService.check_current_user_permissions(item, current_user)
+        return item
 
     @staticmethod
     def create_item(
-        session: SessionDep, current_user: CurrentUser, item_in: ItemCreate
+        session: Session, current_user: User, item_in: ItemCreate
     ) -> Any:
         """
         Create new item.
@@ -69,17 +77,18 @@ class ItemsService:
 
     @staticmethod
     def update_item(
-        session: SessionDep,
-        current_user: CurrentUser,
+        session: Session,
+        current_user: User,
         id: uuid.UUID,
         item_in: ItemUpdate,
     ) -> Any:
         """
         Update an item.
         """
-        item = ItemsRepository.get_user_item_by_id(
+        item: Item = ItemsRepository.get_item_by_id(
             session, current_user, id,
         )
+        ItemsValidationService.check_current_user_permissions(item, current_user)
         update_dict = item_in.model_dump(exclude_unset=True)
         item.sqlmodel_update(update_dict)
         session.add(item)
@@ -90,14 +99,15 @@ class ItemsService:
 
     @staticmethod
     def delete_item(
-        session: SessionDep, current_user: CurrentUser, id: uuid.UUID
+        session: Session, current_user: User, id: uuid.UUID
     ) -> Message:
         """
         Delete an item.
         """
-        item = ItemsRepository.get_user_item_by_id(
+        item = ItemsRepository.get_item_by_id(
             session, current_user, id,
         )
+        ItemsValidationService.check_current_user_permissions(item, current_user)
         session.delete(item)
         session.commit()
         return Message(message="Item deleted successfully")
